Another 230 lines of boilerplate killed
As well as one of the most annoying parts of find-replace work that had to be done.

I'm kind of fine with the use of the macro here instead of wrapping up the parameters in a class - a class would add noise to all of the yaku code by forcing a layer of digging into an object, which feels like a pretty costly price to pay just for not needing to edit 40 function parameters at once.