Refactor tile data structure
'TileKind' replaces the prior union structure 'TileData'. It combines suits and face values, supporting honours too.

'TileInstance' replaces the prior class 'Tile', making it explicit that this class has the instance ID associated with it.

Between the two, a class 'Tile' has been added, which extends TileKind with 'properties' which are rule specifiable. Initially, the akadora property has been defined and included in the standard yonma ruleset.

The result of the refactor is two main benefits:
- code working with tiles is generally simpler as there is no longer a need to convert between several different types
- it's much much clearer how comparisons/sorts are made of tiles, and when we have instance data and when we don't. in particular, the three new types simply implement operator== each rather than needing a dubiously distinct 'operator==, Equvalent, Is' set of three functions.

Introducing the tile properties also resolves a TODO